libdlm: fix udev rules to deal with misc contrains
new kernels do not allow renaming of certain class of char devices,
and that means that new dlm* devices are created in /dev instead
of /dev/misc.

Change udev rules to create compat SYMLINKS in /dev/misc and
maintain the same file modes.

Also update to deal with dlm_plock and drop dlm_default that is
not referenced neither in kernel or userland.

This also fixes a dlm_controld segfaults on closing connections
to dlm_plock.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>