dlm_controld: fix handling of startup partition merge
Nack messages are not matched correctly with previous change
structs.  Normal start messages should not be matched with
changes that preceeded the node's join time, but nack messages
should, so the match_change() function needs to distinguish.

This can occur when two nodes are joining a lockspace, but
neither has completed when a cluster partition+merge happens.

(copying fix from cluster.git
70b15e7afa730166ef0a3c81948d31ab1dd5aa93)

Signed-off-by: David Teigland <teigland@redhat.com>