dlm: clear out addrs before calling into corosync_cft_get_node_addrs()
The corosync_cfg_get_node_addrs() call does not fill the whole of the
addrs field passed in, specifically it only writes the the address
family and IP address, leaving the port number untouched.

If the port number contains junk, then that can get passed into the
kernel by dlm_controld where it is subsequently used in the comparison
that checks for valid cluster nodes in a connection. If this happens
then an otherwise valid connection can be rejected and the dlm will
hang.

I've seen this quite often on s390 but I don't see any reason why it
might not also be causing intermittent connection problems on other
archs.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>